Move Auth0Request into lib.rs
